29company

+++ src/main/java/com/sunvua/ljy/model/UserModel.java
@@ -39,7 +39,7 @@ public class UserModel {
         return gender;
     }
 
-    public void setGender(Byte gender) {
+    public void setGender(Integer gender) {
         this.gender = gender;
     }
